+++ src/test/resources/fixtures/sqs/tier-calculation-event.json
@@ -1,33 +1,33 @@
 {
-    "Type": "Notification",
-    "MessageId": "042e7973-ca2d-551d-916b-21f1f9eb3a54",
-    "TopicArn": "arn:aws:sns:eu-west-2:754256621582:cloud-platform-Digital-Prison-Services-453cac1179377186788c5fcd12525870",
-    "Message": "{\"offenderId\":2500468261,\"crn\":\"X373878\",\"sourceId\":11174,\"eventDatetime\":\"2021-01-20T13:34:59\"}",
-    "Timestamp": "2021-01-20T13:35:02.345Z",
-    "SignatureVersion": "1",
-    "Signature": "wLZFFnngE0vRouyQ6ff/IUQiZOxWitcJ/dJUISFo4mjKMAoz/FDUJW1H31UST9LJLXz1X0h6x4rotFTNK59TNZBnhqjE0+wM1ttaGhVVahJVhZ2GKfOIwjhOjjueK0s9L43Ki+U6nTIVx7LGN/eTQKS/guuNLESY4Wzq2exLNtc5Hr/IMEYGZqAV9ZXvBPbslAqSrkHSzrYATgmLKXDte3E6qh+jsihTgmbpeaR8YQklJklLYeTI2X0y3oteSUqyyDeKLF/I7k/gYDvqTj+9Qme0kCPc7uxZMegFGvgonf5Ln1ylNK99AIW1DssvgpuXRMCrTqZqpkWScM4bDjAPsQ==",
-    "SigningCertURL": "https://sns.eu-west-2.amazonaws.com/SimpleNotificationService-010a507c1833636cd94bdb98bd93083a.pem",
-    "UnsubscribeURL": "https://sns.eu-west-2.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=arn:aws:sns:eu-west-2:754256621582:cloud-platform-Digital-Prison-Services-453cac1179377186788c5fcd12525870:94794e45-0011-44ed-8184-9f50c1858e78",
-    "MessageAttributes": {
-      "eventType": {
-        "Type": "String",
-        "Value": "OFFENDER_MANAGEMENT_TIER_CALCULATION_REQUIRED"
-      },
-      "source": {
-        "Type": "String",
-        "Value": "delius"
-      },
-      "id": {
-        "Type": "String",
-        "Value": "fcf89ef7-f6e8-ee95-326f-8ce87d3b8ea0"
-      },
-      "contentType": {
-        "Type": "String",
-        "Value": "text/plain;charset=UTF-8"
-      },
-      "timestamp": {
-        "Type": "Number.java.lang.Long",
-        "Value": "1611149702333"
-      }
+  "Type": "Notification",
+  "MessageId": "042e7973-ca2d-551d-916b-21f1f9eb3a54",
+  "TopicArn": "arn:aws:sns:eu-west-2:754256621582:cloud-platform-Digital-Prison-Services-453cac1179377186788c5fcd12525870",
+  "Message": "{\"offenderId\":2500468261,\"crn\":\"X373878\",\"sourceId\":11174,\"eventDatetime\":\"2021-01-20T13:34:59\"}",
+  "Timestamp": "2021-01-20T13:35:02.345Z",
+  "SignatureVersion": "1",
+  "Signature": "wLZFFnngE0vRouyQ6ff/IUQiZOxWitcJ/dJUISFo4mjKMAoz/FDUJW1H31UST9LJLXz1X0h6x4rotFTNK59TNZBnhqjE0+wM1ttaGhVVahJVhZ2GKfOIwjhOjjueK0s9L43Ki+U6nTIVx7LGN/eTQKS/guuNLESY4Wzq2exLNtc5Hr/IMEYGZqAV9ZXvBPbslAqSrkHSzrYATgmLKXDte3E6qh+jsihTgmbpeaR8YQklJklLYeTI2X0y3oteSUqyyDeKLF/I7k/gYDvqTj+9Qme0kCPc7uxZMegFGvgonf5Ln1ylNK99AIW1DssvgpuXRMCrTqZqpkWScM4bDjAPsQ==",
+  "SigningCertURL": "https://sns.eu-west-2.amazonaws.com/SimpleNotificationService-010a507c1833636cd94bdb98bd93083a.pem",
+  "UnsubscribeURL": "https://sns.eu-west-2.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=arn:aws:sns:eu-west-2:754256621582:cloud-platform-Digital-Prison-Services-453cac1179377186788c5fcd12525870:94794e45-0011-44ed-8184-9f50c1858e78",
+  "MessageAttributes": {
+    "eventType": {
+      "Type": "String",
+      "Value": "OFFENDER_MANAGEMENT_TIER_CALCULATION_REQUIRED"
+    },
+    "source": {
+      "Type": "String",
+      "Value": "delius"
+    },
+    "id": {
+      "Type": "String",
+      "Value": "fcf89ef7-f6e8-ee95-326f-8ce87d3b8ea0"
+    },
+    "contentType": {
+      "Type": "String",
+      "Value": "text/plain;charset=UTF-8"
+    },
+    "timestamp": {
+      "Type": "Number.java.lang.Long",
+      "Value": "1611149702333"
     }
   }
+}
